Implementations updated

+++ src/forum/server/exceptions/message/MessageNotFoundException.java
@@ -11,7 +11,7 @@ public class MessageNotFoundException extends Exception {
 
 	private static final long serialVersionUID = 1000753555749212018L;
 
-	public MessageNotFoundException(int messageID) {
+	public MessageNotFoundException(long messageID) {
 		super("A message with an id " + messageID + " was not found!");
 	}
 }
